+++ src/components/progress-bar/progress-bar.styles.ts
@@ -6,9 +6,9 @@ export default css`
 
   :host {
     --height: 16px;
-    --track-color: rgb(var(--sl-color-gray-200));
+    --track-color: rgb(var(--sl-color-neutral-500) / 20%);
     --indicator-color: rgb(var(--sl-color-primary-500));
-    --label-color: rgb(var(--sl-color-empty));
+    --label-color: rgb(var(--sl-color-1000));
 
     display: block;
   }