+++ src/components/progress-ring/progress-ring.styles.ts
@@ -5,7 +5,7 @@ export default css`
   ${componentStyles}
 
   :host {
-    --track-color: rgb(var(--sl-color-gray-200));
+    --track-color: rgb(var(--sl-color-neutral-500) / 15%);
     --indicator-color: rgb(var(--sl-color-primary-500));
 
     display: inline-flex;