+++ src/components/spinner/spinner.styles.ts
@@ -5,7 +5,7 @@ export default css`
   ${componentStyles}
 
   :host {
-    --track-color: rgb(var(--sl-color-full) / 10%);
+    --track-color: rgb(var(--sl-color-neutral-500) / 15%);
     --indicator-color: rgb(var(--sl-color-primary-500));
     --stroke-width: 2px;