+++ docs/resources/changelog.md
@@ -13,7 +13,7 @@ This release adds a variety of new color primitives and changes the way color to
 ```css
 .example {
   /* rgb() is required now */
-  color: rgb(var(--sl-color-gray-500));
+  color: rgb(var(--sl-color-neutral-500));
 }
 ```
 
@@ -22,15 +22,17 @@ This is more verbose than before, but it has the advantage of letting you set th
 ```css
 .example-with-alpha {
   /* easily adjust opacity for any color token */
-  color: rgb(var(--sl-color-gray-500) / 50%);
+  color: rgb(var(--sl-color-neutral-500) / 50%);
 }
 ```
 
 This change applies to all design tokens that implement a color. Refer to the [color tokens](/tokens/color) page for more details.
 
 - 🚨 BREAKING: all design tokens that implement colors have been converted to `R G B` and must be used with the `rgb()` function
-- 🚨 BREAKING: removed `--sl-color-black|white` color tokens (use `--sl-color-empty|full` instead)
+- 🚨 BREAKING: removed `--sl-color-black|white` color tokens (use `--sl-color-1000|full` instead)
 - 🚨 BREAKING: removed `--sl-color-primary|success|warning|info|danger-text` design tokens (use theme or primitive colors instead)
+- 🚨 BREAKING: removed `info` variant from `sl-alert`, `sl-badge`, `sl-button`, and `sl-tag` (use `neutral` instead)
+- 🚨 BREAKING: removed `--sl-color-info-*` design token (use `--sl-color-neutral-*` instead)
 - Added new color primitives to the base set of design tokens
 - Added `--sl-color-*-950` swatches to all color palettes
 - Added a console error that appears when menu items have duplicate values in `sl-select`